273: Added SidecarSpec
