+++ src/dependency-manager/src/spec/transform/common-transform.ts
@@ -1,0 +1,30 @@
+import { parse as shell_parse } from 'shell-quote';
+import { LivenessProbeConfig } from '../../config/common-config';
+import { Dictionary } from '../../utils/dictionary';
+import { LivenessProbeSpec } from '../common-spec';
+
+export const transformLivenessProbeSpecCommand = function (command: string[] | string | undefined): string[] | undefined {
+  if (!command) {
+    return undefined;
+  }
+  if (typeof command === 'string') {
+    return shell_parse(command.replace(/\$/g, '__arc__')).map(e => `${e}`.replace(/__arc__/g, '$'));
+  } else {
+    return command;
+  }
+};
+
+export const transformLivenessProbeSpec = function (liveness_probe: LivenessProbeSpec | undefined, environment: Dictionary<string | null>): LivenessProbeConfig | undefined {
+  if (!liveness_probe || !Object.keys(liveness_probe).length) { return undefined; }
+
+  return {
+    success_threshold: liveness_probe.success_threshold || LivenessProbeSpec.default_success_threshold,
+    failure_threshold: liveness_probe.failure_threshold || LivenessProbeSpec.default_failure_threshold,
+    timeout: liveness_probe.timeout || LivenessProbeSpec.default_timeout,
+    interval: liveness_probe.interval || LivenessProbeSpec.default_interval,
+    initial_delay: liveness_probe.initial_delay || LivenessProbeSpec.default_initial_delay,
+    path: liveness_probe.path,
+    command: transformLivenessProbeSpecCommand(liveness_probe.command),
+    port: liveness_probe.port,
+  };
+};
