+++ src/dependency-manager/src/spec/transform/component-transform.ts
@@ -1,10 +1,12 @@
 import { ComponentConfig, ComponentInstanceMetadata, ComponentInterfaceConfig, OutputDefinitionConfig, ParameterDefinitionConfig } from '../../config/component-config';
-import { ComponentContext, OutputValue, ParameterValue, ServiceContext, TaskContext } from '../../config/component-context';
+import { ComponentContext, OutputValue, ParameterValue, ServiceContext, SidecarContext, TaskContext } from '../../config/component-context';
 import { ServiceInterfaceConfig } from '../../config/service-config';
 import { Dictionary, transformDictionary } from '../../utils/dictionary';
 import { ComponentInterfaceSpec, ComponentSpec, OutputDefinitionSpec, ParameterDefinitionSpec } from '../component-spec';
+import { ServiceSpec } from '../service-spec';
 import { ComponentSlug, ComponentSlugUtils, Slugs } from '../utils/slugs';
 import { transformServiceSpec } from './service-transform';
+import { transformSidecarSpec } from './sidecar-transform';
 import { transformTaskSpec } from './task-transform';
 
 export const transformComponentSpecName = (name: string): ComponentSlug => {
@@ -124,6 +126,13 @@ export const transformComponentContext = (
     };
   }
 
+  const sidecar_context: Dictionary<SidecarContext> = {};
+  for (const [sk, sv] of Object.entries(config.sidecars)) {
+    sidecar_context[sk] = {
+      environment: sv.environment,
+    };
+  }
+
   const task_context: Dictionary<TaskContext> = {};
   for (const [tk, tv] of Object.entries(config.tasks)) {
     task_context[tk] = {
@@ -139,14 +148,26 @@ export const transformComponentContext = (
     ingresses: ingress_context,
     interfaces: interface_context,
     services: service_context,
+    sidecars: sidecar_context,
     tasks: task_context,
   };
 };
 
 export const transformComponentSpec = (spec: ComponentSpec, source_yml: string, tag: string, instance_metadata?: ComponentInstanceMetadata): ComponentConfig => {
+  // Inject the component-level sidecars into each of the component service specs
+  let spec_services = spec.services;
+  if (spec_services) {
+    spec_services = Object.entries(spec_services).reduce((acc, [service_name, service_spec]) => {
+      const merged_sidecars = { ...spec.sidecars, ...service_spec.sidecars };
+      const merged_service = { ...service_spec, sidecars: merged_sidecars };
+      return { ...acc, [service_name]: merged_service };
+    }, {} as Dictionary<ServiceSpec>);
+  }
+
   const parameters = transformDictionary(transformParameterDefinitionSpec, spec.parameters);
   const outputs = transformDictionary(transformOutputDefinitionSpec, spec.outputs);
-  const services = transformDictionary(transformServiceSpec, spec.services, spec.name, tag, instance_metadata);
+  const services = transformDictionary(transformServiceSpec, spec_services, spec.name, tag, instance_metadata);
+  const sidecars = transformDictionary(transformSidecarSpec, spec.sidecars, spec.name, tag, instance_metadata);
   const tasks = transformDictionary(transformTaskSpec, spec.tasks, spec.name, tag, instance_metadata);
   const interfaces = transformDictionary(transformComponentInterfaceSpec, spec.interfaces);
   const dependencies = spec.dependencies || {};
@@ -168,6 +189,7 @@ export const transformComponentSpec = (spec: ComponentSpec, source_yml: string, 
     outputs,
 
     services,
+    sidecars,
     tasks,
 
     dependencies,
@@ -185,6 +207,7 @@ export const transformComponentSpec = (spec: ComponentSpec, source_yml: string, 
         outputs,
         interfaces,
         services,
+        sidecars,
         tasks,
       } as ComponentConfig
     ),
